pyproject.toml
[project]
name = "rl-llm-autotrainer"
version = "0.1.0"
description = ""
authors = [
    {name = "charitarthchugh",email = "37895518+charitarthchugh@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13,<3.15"
dependencies = [
    "torch (>=2.9.1,<3.0.0)",
    "transformers (>=4.35.0)",
    "peft (>=0.6.0)",
    "datasets (>=2.14.0)",
    "accelerate (>=0.24.0)",
    "bitsandbytes (>=0.41.0)",
    "pandas (>=1.5.0)",
    "numpy (>=1.24.0)",
    "scikit-learn (>=1.3.0)",
    "evaluate (>=0.4.0)",
    "python-dotenv (>=1.0.0)",
    "huggingface-hub (>=0.17.0)",
    "sentencepiece (>=0.1.99)",
    "protobuf (>=3.20.0)",
    "absl-py (>=2.3.1,<3.0.0)",
    "nltk (>=3.9.2,<4.0.0)",
    "rouge-score (>=0.1.2,<0.2.0)",
    "bert-score (>=0.3.13,<0.4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
